feat(frost-signer): carve persistence into seams
Introduces StateKeyProvider trait + EnvKeyProvider / CommandKeyProvider /
CachedStateKeyProvider struct adapters (replacing the prior enum-based
dispatch), then splits persistence.rs into 4 pub(crate) submodules:
schema_codec, envelope_io, key_provider, pending_ops. mod.rs
re-exports so call sites are unchanged. Persisted schemas
(PersistedSessionState, PersistedEngineState, PersistedKeyPackage)
byte-for-byte stable. TryFrom impl bodies unchanged.
